requirements.txt
google-ads>=23.0.0
mcp>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0.0
pytest>=9.0.0
